requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
configparser==5.0.0
discord.py==1.4.1
ffmpeg==1.4
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
praw==7.1.0
prawcore==1.5.0
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
toml==0.10.1
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
wrapt==1.12.1
yarl==1.5.1
